requirements.txt
tf_grpc_plugin
msgpack
requests
nltk
pyyaml
click
